[CL-1334] Rename files
