--- front/app/containers/Admin/formBuilder/components/Tab.tsx
@@ -1,32 +1,0 @@
-import { FC, useEffect } from 'react';
-
-import { InjectedIntlProps } from 'react-intl';
-import { InsertConfigurationOptions, ITab } from 'typings';
-import { injectIntl } from 'utils/cl-intl';
-
-import messages from './messages';
-
-interface Props {
-  onData: (data: InsertConfigurationOptions<ITab>) => void;
-}
-
-const Tab: FC<Props & InjectedIntlProps> = ({
-  onData,
-  intl: { formatMessage },
-}) => {
-  useEffect(() => {
-    onData({
-      configuration: {
-        label: formatMessage(messages.surveyTitle),
-        url: 'native-survey',
-        feature: 'native_surveys',
-        name: 'nativeSurveys',
-      },
-      insertAfterName: 'survey-results',
-    });
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-  return null;
-};
-
-export default injectIntl(Tab);
